Resolve ambiguous scalar operator overloads
The earlier attempt to resolve this problem in 9286bf4 failed for two
reasons:

1. Static asserts were checking whether the argument was a unit type
instead of a unit value type, all arguments met this condition.

2. Static asserts do not participate in SFINAE because they are not
evaluated in the immediate context of the function type or its template
parameter types.

Adding an appropriate metafunction and using it in an enable_if
expression in the template parameter list removes these overloads from
the candidate set as intended.