Replace static asserts with enable_if
This commit continues the changes in 9fdd9cf by replacing remaining
static asserts with equivalent enable_if expressions. This will cause
errors to be emitted at the callsite of the invalid operation rather
than at the location of the static_assert which required additional
tracing to find the source of the error.

Many of the enable_if expressions use a pattern where the test condition
uses a transient template variable (e.g. typename _Tw = _Ty) in order to
force deferred evaluation of the expression. This is necessary to avoid
errors due to implicit instantiation of undefined templates (the failure
condition used for SFINAE) during class instantiation.

Additional minor changes:

* Removed unnecessary divide delegate since division by same type now
  returns a dimensionless value type rather than scalar type (6dd8d04)

* Added mpl::units::root metafunction to simplify expressions using
  mpl::units::detail::fn_root (cf. sqrt, cbrt)

* Removed operator^ and operator%. These were intended for use with
  cross or exterior product operators for vector types but dimensional
  correctness is difficult to guarantee and there are other more serious
  issues preventing general usage with vector types.